+++ Andromeda/Input/Vita/GamepadDeviceVita.h
@@ -19,11 +19,11 @@ namespace Andromeda
 
 		public:
 
-			int LeftAnalogX();
-			int LeftAnalogY();
+			float LeftAnalogX();
+			float LeftAnalogY();
 
-			int RightAnalogX();
-			int RightAnalogY();
+			float RightAnalogX();
+			float RightAnalogY();
 
 			bool KeyDown(Gamepad::Button button);
 			bool KeyUp(Gamepad::Button button);
